chore

+++ scripts/prerender.ts
@@ -3,7 +3,7 @@ import fs from 'node:fs'
 import path from 'node:path'
 
 /**
- * 预渲染首页
+ * https://cn.vitejs.dev/guide/ssr.html#pre-rendering--ssg
  */
 export default (options?: {
   routes?: string[]
@@ -22,12 +22,15 @@ export default (options?: {
     async writeBundle() {
       const routesToPrerender = options?.routes || ['/']
       const ssrDist = config.build.outDir
-      const { render } = await import(path.resolve(ssrDist, 'server.mjs'))
+      const ssrEntry = path.resolve(ssrDist, 'server.mjs')
+      if (!fs.existsSync(ssrEntry)) return
+      const { render } = await import(ssrEntry)
       const clientDist = options?.clientDist || path.resolve(config.root, 'dist')
       const template = fs.readFileSync(path.resolve(clientDist, 'index.html'), 'utf-8')
+      const defaultReplaceMark = /<!--ssr-start-->([\s\S]+)<!--ssr-end-->/
       for (const url of routesToPrerender) {
         const html = await render(url)
-        const replaceMark = options?.replaceMark || /<!--ssr-start-->([\s\S]+)<!--ssr-end-->/
+        const replaceMark = options?.replaceMark || defaultReplaceMark
         const renderedHtml = template.replace(replaceMark, html)
         const filePath = url === '/' ? 'index.html' : url.endsWith('.html') ? url : url + '.html'
         fs.writeFileSync(path.resolve(clientDist, filePath), renderedHtml)
